tarogarden: run tests serially as a temporary measure
As is, we run into a race condition when trying to initialize both
sqlite db instances. Oddly, the `tarodb` tests work just fine, and each
of them makes a new instance of a db.

I attempted to update to the latest versionw which has a fix here:
https://gitlab.com/cznic/sqlite/-/issues/77. However it runs into
another race issue elsewhere.